FABN-1519: Refactor of full and private event factory
Simplification of factory code for creating private events.
